HW2NOOPTIONAL

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryUserMealRepositoryImpl.java
@@ -1,11 +1,12 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.model.UserMeal;
 import ru.javawebinar.topjava.repository.UserMealRepository;
 import ru.javawebinar.topjava.util.UserMealsUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -21,7 +22,7 @@ public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
         UserMealsUtil.MEAL_LIST.forEach(this::save);
     }
 
-    @Override
+
     public UserMeal save(UserMeal userMeal) {
         if (userMeal.isNew()) {
             userMeal.setId(counter.incrementAndGet());
@@ -31,18 +32,53 @@ public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id) {
+        boolean result = false;
+
+        for( Map.Entry<Integer, UserMeal> entry : repository.entrySet() ) {
+            if (id == (entry.getKey()))
+                repository.remove(id);
+                result = true;
+        }
+        return result;
     }
 
     @Override
     public UserMeal get(int id) {
-        return repository.get(id);
+        int result = 0;
+
+        for( Map.Entry<Integer, UserMeal> entry : repository.entrySet() ) {
+            if (id == (entry.getKey()))
+                result = id;
+        }
+        if(result == 0)
+            return repository.get(result);
+        else
+        return null;
     }
 
     @Override
     public Collection<UserMeal> getAll() {
-        return repository.values();
+
+        Comparator comparator = new SortedByDate();
+        List<UserMeal> mealList = new ArrayList<>(repository.values());
+
+        Collections.sort(mealList,comparator);
+        Collections.reverse(mealList);
+        return mealList;
+    }
+
+    class SortedByDate implements Comparator<UserMeal> {
+
+        public int compare(UserMeal obj1, UserMeal obj2) {
+
+            String str1 = obj1.getDateTime().toString();
+            String str2 = obj2.getDateTime().toString();
+
+            return str1.compareTo(str2);
+        }
+
+
     }
 }
